refactor: separate editing logic
