+++ src/hook/usePart/usePart.tsx
@@ -2,16 +2,12 @@ import { BASE_URL } from "@/constants";
 import UserContext from "@/context/userContext";
 import { PartType } from "@/schemas/parts";
 import { notifyPositionMap, notifyType } from "@/types";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useQueryClient } from "@tanstack/react-query";
 import axios, { AxiosError } from "axios";
 import { useContext } from "react";
 import useNotify from "../useNotify";
 
-type UsePartProps = {
-  partId?: number;
-};
-
-const usePart = ({ partId }: UsePartProps) => {
+const usePart = () => {
   const queryClient = useQueryClient();
 
   const {
@@ -50,37 +46,7 @@ const usePart = ({ partId }: UsePartProps) => {
     }
   };
 
-  const urlEditPart = `${BASE_URL}/part/${partId}`;
-
-  const editPartMutation = useMutation({
-    mutationFn: (formData: PartType) => {
-      const payload = {
-        ...formData,
-        price: Number(formData.price),
-        quantity: Number(formData.quantity),
-      };
-      return axios.put(urlEditPart, payload, header);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["getAllPart"] });
-      notify(
-        "Peça Editado com Sucesso!",
-        notifyPositionMap.topRight,
-        notifyType.success
-      );
-    },
-    onError: (error) => {
-      const err = error as AxiosError;
-
-      notify(
-        err.message as string,
-        notifyPositionMap.topRight,
-        notifyType.error
-      );
-    },
-  });
-
-  return { registerPart, editPartMutation };
+  return { registerPart };
 };
 
 export default usePart;
